Module easy_plugin_parsers::arguments

Arguments.

Structs

Arguments
A set of parsed arguments.

EnumArguments
A set of parsed arguments found in an enum.

SequenceArguments
A set of parsed arguments found in a sequence.

Functions

parse_arguments
Parses the supplied arguments with the supplied argument specification.